// components/account/type.ts
export interface ProfileDetails
{
    email: string;
    companyName: string;
    gstNumber: string;
    memberSince: string;
    status: string;
    addressLine1: string;
    addressLine2: string;
    mobileNo: string;
    country: string;
    city: string;
    state: string;
    district: string;
    pincode: string;
    walletBalance: number
}